Writing structs in a binary file - Server side
Tried writing individual structs and then overwrite its content in the binary file, but the seekg of the fstream library was not functioning like it supposed to. Then, I thought about a second option and that is: writing an array of structs in the binary file, then change the array (out of the file) and then overwrite the binary file so that it only contains the changed array.